% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineCSV.R
\name{combineCSV}
\alias{combineCSV}
\title{Combine boundary extraction CSV files}
\usage{
combineCSV(csv_dir)
}
\arguments{
\item{csv_dir}{Character. Directory containing watershed CSV files.}
}
\value{
A combined data.frame sorted by datetime.
}
\description{
Reads all CSV files produced by the boundary extraction and combines them
into a single data frame sorted by datetime.
Special case: files ending with "0_0.csv" will have "00:00:00" added to datetime if missing.
}
