% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullMRMS.R
\name{pullMRMS}
\alias{pullMRMS}
\title{Download MRMS data}
\usage{
pullMRMS(start, end, destination, product)
}
\arguments{
\item{start}{DateTime of the first timestep}

\item{end}{DateTime of the last timestep}

\item{destination}{Destination where the data are stored}

\item{product}{MRMS product (RQI, MultiSensor QPE, Radar only QPE, SPR)}
}
\value{
Returns any missing dates found during download
}
\description{
Download MRMS data
}
\examples{
pullMRMS(start = ymd_hm("2025-07-03 20:00"),
  end = ymd_hm("2025-07-04 23:00"),
  destination = './NFork_GuadRiv/mrms_rasters',
  product = "SurfacePrecipRate")
}
